+++ noodles-cram/src/crai.rs
@@ -10,7 +10,7 @@ mod writer;
 pub use self::{reader::Reader, record::Record, writer::Writer};
 
 #[cfg(feature = "async")]
-pub use self::r#async::Reader as AsyncReader;
+pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};
 
 use std::{fs::File, io, path::Path};