+++ noodles-cram/src/crai/async.rs
@@ -1,8 +1,9 @@
 //! Async CRAM index I/O.
 
 mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};
 
 use std::path::Path;